Corrigido a parte do cpf, que estava com int

+++ Genesys/src/main/java/DAO/UsuarioDAO.java
@@ -10,7 +10,7 @@ package DAO;
 		
 		public void remover(Usuario usuario);
 		
-		public void pesquisar(int id_cpf);
+		public void pesquisar(double id_cpf);
 		
 		public List<Usuario> listaUsuario();
